requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-ninja==0.20.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
idna==3.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
